Math.Chebyshev: Fix ambiguous calls to pow with gcc-7

+++ include/boost/math/special_functions/chebyshev.hpp
@@ -44,7 +44,7 @@ inline Real chebyshev_imp(unsigned n, Real const & x)
         if (x > 1 || x < -1)
         {
             Real t = sqrt(x*x -1);
-            return static_cast<Real>((pow(x+t, n+1) - pow(x-t, n+1))/(2*t));
+            return static_cast<Real>((pow(x+t, (int)(n+1)) - pow(x-t, (int)(n+1)))/(2*t));
         }
         T1 = 2*x;
     }
